+++ api/accounts/models.py
@@ -1,6 +1,6 @@
+from authemail.models import EmailUserManager, EmailAbstractUser
 from django.core.validators import RegexValidator
 from django.db import models
-from authemail.models import EmailUserManager, EmailAbstractUser
 from django.utils.translation import gettext_lazy as _
 
 
@@ -13,6 +13,8 @@ class MyUser(EmailAbstractUser):
                                          "'+999999999'. Up to 15 digits allowed.")
     phone = models.CharField(validators=[phone_regex], max_length=17, blank=False,
                              help_text=_("Phone number, preferrably mobile"), verbose_name=_("Phone number"))
+    receive_notifications = models.BooleanField(default=False, verbose_name=_("Receive Signup Notifications"),
+                                                help_text=_("Receive a Email Notification when a new team is created"))
 
     # Required
     objects = EmailUserManager()